python/requirements.txt
flask
python-dotenv
groq
requests
